dom/tests/mochitest/bugs/test_DOMWindowCreated_chromeonly.html

<!DOCTYPE html>
<head>
  <title>DOMWindowCreated not visible in content</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<body onload="ok(true, 'Test finished'); SimpleTest.finish();">
  <p id="display"></p>

  <script type="application/javascript">
  SimpleTest.waitForExplicitFinish();
  window.addEventListener("DOMWindowCreated", function() { ok(false, "DOMWindowCreated should not have fired"); }, false);

  <iframe src="data:text/plain,Hi"></iframe>
